Extract each matcher's tests to its own test file
