+++ setupTests.js
@@ -1,0 +1,4 @@
+import {plugins} from 'pretty-format'
+import './src/extend-expect'
+
+expect.addSnapshotSerializer(plugins.ConvertAnsi)
